+++ apps/docs/app/pages/index.vue
@@ -1,0 +1,19 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div class="space-y-4 p-4">
+    <div class="space-x-4">
+      <NuxtLink to="/getting-started/intro">
+        <MButton label="Intro" />
+      </NuxtLink>
+      <NuxtLink to="/getting-started/installation">
+        <MButton label="Installation" />
+      </NuxtLink>
+    </div>
+  </div>
+</template>
+
+<style scoped>
+
+</style>