+++ packages/nuxt/src/runtime/components/prose/callout/Danger.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+import Callout from '../Callout.vue'
+import { useAppConfig } from '#imports'
+
+const appConfig = useAppConfig()
+</script>
+
+<template>
+  <Callout color="danger" :icon="appConfig.mockline.icons.danger">
+    <slot mdc-unwrap="p" />
+  </Callout>
+</template>