+++ packages/nuxt/src/runtime/utils/useComponent.ts
@@ -1,6 +1,6 @@
 import { computed, type MaybeRefOrGetter, toValue } from 'vue'
 import { twMerge } from 'tailwind-merge'
-import * as components from '@mockline/themes'
+import { components } from '@mockline/themes'
 import { useAppConfig } from '#imports'
 
 /**
@@ -55,7 +55,7 @@ export function useComponent<
   const ui = computed(() => {
     const baseComponent = components[componentName]
     const resolvedProps = toValue(componentProps) || {}
-    return baseComponent(resolvedProps)
+    return baseComponent({ ...resolvedProps, transitions: appConfig.mockline?.transitions })
   })
 
   return {