+++ packages/themes/src/components/index.ts
@@ -1,0 +1,14 @@
+// Components
+export * from './button'
+export * from './sidebar'
+
+// Prose
+export * from './prose/list'
+export * from './prose/code'
+export * from './prose/pre'
+export * from './prose/codeGroup'
+export * from './prose/callout'
+
+// Typography
+export * from './prose/typography'
+